addHeadLine is now a function which is easier to compose
