+++ csv2xlslib.Test/funcompTest.cpp
@@ -9,6 +9,9 @@ namespace csv2xls
 {
 namespace funcompTest
 {
+    using funcomp::operator|;
+    using funcomp::repeatUntil;
+
     struct Tracer{
         Tracer(){std::cout << "Tracer()\n";}
         Tracer(Tracer const& ){std::cout << "Tracer(Tracer const&)\n";}
@@ -20,91 +23,42 @@ namespace funcompTest
 
     struct Data
     {
-        std::stringstream file;
+        int file;
         Tracer tracer;
     };
 
-    template<typename Callee, typename Predicate>
-    struct helper
-    {
-        Callee callee;
-        Predicate pred;
-        helper(Callee callee_, Predicate pred_)
-        :callee(callee_), pred(pred_)
-        {
-        }
-        template<typename... ARGs>
-        auto operator()(ARGs&&... args)const
-        {
-            while (true)
-            {
-                auto const result = callee(std::forward<ARGs>(args)...);
-                if (pred(result))
-                    return result;
-            }
 
+    auto innerFunc1 = [](Data data)->std::optional<Data>{
+        if (data.file == 3)
+        {
+            return data;
         }
+        return std::nullopt;
     };
 
-    template <typename Callee, typename Caller> auto operator>>=(Callee callee, Caller caller)
+    auto innerFunc2(int i)
     {
-        if constexpr (std::is_base_of_v<funcomp::HigherOrderFunction, Caller>)
-        {
-            return [hof = caller(callee)](auto&&... arg) { return hof(arg...); };
-        }
-        else
-        {
-            return [callee, caller](auto&&... arg) { return caller(callee(arg...)); };
-        }
-    }
-
-    template <typename Pred> struct RepeatUntil : public funcomp::HigherOrderFunction
-    {
-        Pred m_pred;
-        RepeatUntil(Pred pred)
-                : m_pred(pred)
-        {
-        }
-        template <typename Callee> auto operator()(Callee callee) const
-        {
-            return helper<Callee, Pred>(callee,m_pred);
-        }
-    };
-
-    auto innerFunc1 = [](std::stringstream & stream)->char{
-        if (stream.good())
-        {
-            char val = static_cast<char>(stream.get());
-            auto count = stream.gcount();
-            if( count == 1 )
-                return val;
-        }
-        return '@';
-    };
-
-    auto innerFunc2(std::vector<char>& out)
-    {
-        return [&](char i)->char{
-            if (i != '@')
-                out.push_back(i);
-            return i;
+        return [&](std::optional<Data> out)->int{
+            if (out->file == 2 && i != '@')
+                return i;
+            return i+1;
 
         };
     }
 
-    auto isEnd = [](char i){
+    auto isEnd = [](int i){
         return i == '@';
     };
 
     TEST_CASE("lol")
     {
-        std::stringstream  file;
-        file << "12345";
-        std::vector<char> out;
-        auto myfun =  RepeatUntil(isEnd)(innerFunc1 >>= innerFunc2(out));
-
-        myfun(file);
-        REQUIRE(out == std::vector<char>{'1','2','3','4','5'});
+//        std::stringstream  file;
+//        file << "12345";
+//        std::vector<char> out;
+//        auto myfun =  innerFunc1|innerFunc2(2)|repeatUntil(isEnd);
+//
+//        myfun(Data{});
+//        REQUIRE(out == std::vector<char>{'1','2','3','4','5'});
     }
 
 }
