+++ csv2xlslib/parseCsvFile.cpp
@@ -17,7 +17,24 @@ namespace csv2xls
             return XlsWorkBook(config.xls_sheet_name);
         };
     }
+    auto addHeadLineFrom(Parameter parameter, std::fstream &input_stream){
+        std::string headline_to_move;
+        if (parameter.input_has_head_line.Get()){
+            std::getline(input_stream, headline_to_move);
+            headline_to_move.append("\n");
+        }
+        return [headline   = std::move(headline_to_move),
+                parameter_ = std::move(parameter)
+               ](OutputDoc output_doc){
+            std::stringstream csv_input(headline);
+            auto buffer = Buffer(parameter_.input_buffer_size);
 
+          return (parameter_.input_has_head_line.Get())
+                    ?convertCsv(buffer, parameter_, csv_input, std::move(output_doc))
+                    :std::optional<OutputDoc>(std::move(output_doc));
+//            return output_doc;
+        };
+    }
     auto makeOutputDoc=[](auto file_gen ) {
             return OutputDoc(file_gen());
     };
@@ -41,11 +58,13 @@ namespace csv2xls
     template<typename NumGen>
     auto writeIntoFile (OutputFileName output_file_name, NumGen num_gen)
     {
-        return [=](OutputDoc output_doc){
-            if(output_doc.isEmpty())
+        return [=](std::optional<OutputDoc> output_doc){
+            if (!output_doc)
+                return WriteStatus::Empty;
+            if(output_doc->isEmpty())
                 return WriteStatus::Empty;
 
-            auto const result = output_doc.writeInto(output_file_name, num_gen());
+            auto const result = output_doc->writeInto(output_file_name, num_gen());
             if(result)
                 return WriteStatus::Error;
 
@@ -57,17 +76,22 @@ namespace csv2xls
     {
         using funcomp::operator|;
         using funcomp::repeatUntil;
+
         if (config.exit_clean) return 0;
 
-        auto const parameter = Parameter{config.output_row_limit, config.csv_separator};
+        auto const parameter = Parameter{config.output_row_limit,
+                                         config.csv_separator,
+                                         config.input_buffer_size,
+                                         config.csv_file_has_headline};
 
         auto csv_input    = openCsvFile(config.csv_file_name);
         auto buffer       = Buffer(config.input_buffer_size);
         auto writeFile    = writeIntoFile(config.output_file_name, EndlessRange(FileNumber(0)));
-        auto convert      = [&](OutputDoc output_doc){
+        auto convert      = [&](std::optional<OutputDoc> output_doc){
                                return convertCsv(buffer, parameter, csv_input, std::move(output_doc));
                             };
-        auto csv2doc      = makeOutputDoc|convert;
+        auto addHeadLine  = addHeadLineFrom(parameter, csv_input);
+        auto csv2doc      = makeOutputDoc|addHeadLine|convert;
         auto isEmptyOrErr = isEitherOf(WriteStatus::Empty, WriteStatus::Error);
         auto convertWith  = csv2doc|writeFile|repeatUntil(isEmptyOrErr);
